% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11janitortest-package.R
\docType{package}
\name{cpp11janitortest-package}
\alias{cpp11janitortest}
\alias{cpp11janitortest-package}
\title{cpp11janitortest: Testing cpp11janitor}
\description{
Testing cpp11janitor. I run this within a tiny package to test the compiled code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/cpp11janitor}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11janitor/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
